refactor: decompose DiaryPreviewGeneratingScreen into focused widgets (Phase 3-2)
Extract DiaryGeneratingSkeletonView and DiaryGeneratingBanner from the
504-line generating screen, reducing it to a layout shell (~200 lines).
Add widget tests for both extracted widgets (10 cases).
